version_05_prd_generator/requirements.txt
# Required packages for PRD Generator
agno
google-generativeai
google-genai
python-dotenv